+++ Package.swift
@@ -16,18 +16,18 @@ let package = Package(
     dependencies: [
         .package(
             url: "https://github.com/sendbird/sendbird-chat-sdk-ios",
-            from: "4.34.1"
+            from: "4.36.0"
         ),
         .package(
             url: "https://github.com/sendbird/sendbird-uikit-ios-spm",
-            from: "3.32.4"
+            from: "3.33.1"
         ),
     ],
     targets: [
         .binaryTarget(
             name: "SendbirdSwiftUI",
-            url: "https://github.com/sendbird/sendbird-swiftui-ios/releases/download/1.1.3/SendbirdSwiftUI.xcframework.zip", // SendbirdSwiftUI_URL
-            checksum: "7d0e6ebf7305cdd7aae183b4810d12ebf5dc440900f8b242de1b2affd3a7c81a" // SendbirdSwiftUI_CHECKSUM
+            url: "https://github.com/sendbird/sendbird-swiftui-ios/releases/download/1.1.4/SendbirdSwiftUI.xcframework.zip", // SendbirdSwiftUI_URL
+            checksum: "79eb08f7b0cf1ad10e9b8322a787033d70be04e2ec78e0c73a958eef6457d057" // SendbirdSwiftUI_CHECKSUM
         ),
         .target(
             name: "SendbirdSwiftUITarget",