+++ CHANGELOG.md
@@ -1,2 +1,4 @@
-## Improvements
-- Fixed missing navigation bar when using SendbirdSwiftUI views in a TabView
+## Changes
+- Updated dependency SDK versions to the latest stable releases
+    - SendbirdChatSDK: 4.36.0
+    - SendbirdMessageTemplate: 3.33.1